wmi provider is buildable
hg-changeset: 31177952f0a8b39bf007e5c353cbdfb4fb6210e2

+++ cimple/wmi/utils.cpp
@@ -68,7 +68,7 @@ String wstr2str(LPWSTR str)
 
     int alen = WideCharToMultiByte(CP_ACP, 0, str, -1, NULL, 0, NULL, NULL);
 
-    if (alen <= sizeof(buf))
+    if (alen <= (int)sizeof(buf))
         astr = buf;
     else
         astr = (char*)::operator new(alen);
